// packages/core/src/middleware/constants/configType.d.ts
/**
 * @experimental  PASS
 */
export declare enum CONFIGTYPE {
    BOXGEOMETRY = "BoxGeometry",
    SPHEREGEOMETRY = "SphereGeometry",
    LOADGEOMETRY = "LoadGeometry",
    PLANEGEOMETRY = "PlaneGeometry",
    CIRCLEGEOMETRY = "CircleGeometry",
    CONEGEOMETRY = "ConeGeometry",
    CYLINDERGEOMETRY = "CylinderGeometry",
    DODECAHEDRONGEOMETRY = "DodecahedronGeometry",
    TORUSGEOMETRY = "TorusGeometry",
    RINGGEOMETRY = "RingGeometry",
    EDGESGEOMETRY = "EdgesGeometry",
    LINECURVEGEOMETRY = "LineCurveGeometry",
    SPLINECURVEGEOMETRY = "SplineCurveGeometry",
    CUBICBEZIERCURVEGEOMETRY = "CubicBezierCurveGeometry",
    QUADRATICBEZIERCURVEGEOMETRY = "QuadraticBezierCurveGeometry",
    CUSTOMGEOMETRY = "CustomGeometry",
    LINETUBEGEOMETRY = "LineTubeGeometry",
    SPLINETUBEGEOMETRY = "SplineTubeGeometry",
    LINESHAPEGEOMETRY = "LineShapeGeometry",
    OBJECT3D = "Object3D",
    MESH = "Mesh",
    LINE = "Line",
    POINTS = "Points",
    SPRITE = "Sprite",
    GROUP = "Group",
    CSS3DOBJECT = "CSS3DObject",
    CSS3DSPRITE = "CSS3DSprite",
    CSS3DPLANE = "CSS3DPlane",
    CSS2DPLANE = "CSS2DPlane",
    IMAGETEXTURE = "ImageTexture",
    CUBETEXTURE = "CubeTexture",
    CANVASTEXTURE = "CanvasTexture",
    VIDEOTEXTURE = "VideoTexture",
    LOADTEXTURE = "LoadTexture",
    MESHBASICMATERIAL = "MeshBasicMaterial",
    MESHSTANDARDMATERIAL = "MeshStandardMaterial",
    MESHPHONGMATERIAL = "MeshPhongMaterial",
    SPRITEMATERIAL = "SpriteMaterial",
    LINEBASICMATERIAL = "LineBasicMaterial",
    POINTSMATERIAL = "PointsMaterial",
    SHADERMATERIAL = "ShaderMaterial",
    MESHPHYSICALMATERIAL = "MeshPhysicalMaterial",
    AMBIENTLIGHT = "AmbientLight",
    SPOTLIGHT = "SpotLight",
    POINTLIGHT = "PointLight",
    DIRECTIONALLIGHT = "DirectionalLight",
    HEMISPHERELIGHT = "HemisphereLight",
    RECTAREALIGHT = "RectAreaLight",
    PERSPECTIVECAMERA = "PerspectiveCamera",
    ORTHOGRAPHICCAMERA = "OrthographicCamera",
    WEBGLRENDERER = "WebGLRenderer",
    CSS3DRENDERER = "CSS3DRenderer",
    SCENE = "Scene",
    TRNASFORMCONTROLS = "TransformControls",
    ORBITCONTROLS = "OrbitControls",
    SMAAPASS = "SMAAPass",
    UNREALBLOOMPASS = "UnrealBloomPass",
    SELECTIVEBLOOMPASS = "SelectiveBloomPass",
    SSAOPASS = "SSAOPass",
    SCRIPTANIMATION = "ScriptAnimation",
    KEYFRAMEANIMATION = "KeyframeAnimation"
}